Restructure entire application
The application has now been changed in two major ways.

1. There is no more tracking of which images are stored locally. That
originally was used to determine if images of disk match those we expect
to have, to prevent accumlating unintend images. However, a simpler way
to do this is to always remove all images in the final processed
directory before moving the new set in.

2. The application can now find images for multiple subreddits. This
will be important in upcoming changes where we add a UI where users can
select subreddits to follow.