Add tray menu item and move image processing to background
There are two major changes in this commit.

1) Adds a menu tray item. This is empty now but will contain the UI for
selecting which subreddits you want to choose.

1) Image fetching and processing is very CPU intensive, and would block
GUI updates. For example, clicking on the tray item would often yield a
spinning volleyball of death. In this commit, we move all of the image
processing to a hidden renderer process, which thus frees up the main
process.